Major refactor in prep for expansion
