+++ .gitignore
@@ -5,3 +5,4 @@ Scripts/__pycache__/*.pyc
 Scripts/server/__pycache__/**
 Scripts/integration/__pycache__/**
 Scripts/utils/__pycache__/**
+.win_venv
