+++ Scripts/client/backend/convert_image_sequence.py
