+++ Scripts/client/slack/slack_api.py
@@ -4,7 +4,7 @@ import json
 
 from pprint import pprint
 
-from server.blocks import SlackBlocks
+from Scripts.slack_components.blocks import SlackBlocks
 
 
 class UploadContent:
