+++ Scripts/client/slack/slack_config.py
